refactor: move test-only schema helpers into tests
Moved applyOpaqueOutputSchemas and envelopeOpaque into tier2_test_helpers_test.go after verifying they were only used by tests, and removed the now-empty production output_schemas.go file. Gates: go build ./..., go vet ./..., go test -count=1 ./..., golangci-lint run, make catalog-drift.
