+++ internal/tools/tier2_test_helpers_test.go
@@ -28,3 +28,23 @@ func tier2Handlers(svc *Service, name string) ([]mcp.ToolDescriptor, bool) {
 	}
 	return applyOpaqueOutputSchemas(normalizeDescriptors(build(svc))), true
 }
+
+func applyOpaqueOutputSchemas(in []mcp.ToolDescriptor) []mcp.ToolDescriptor {
+	for i := range in {
+		if in[i].Tool.OutputSchema == nil {
+			in[i].Tool.OutputSchema = envelopeOpaque(in[i].Tool.Name)
+		}
+	}
+	return in
+}
+
+func envelopeOpaque(action string) map[string]any {
+	return map[string]any{
+		"type":     "object",
+		"required": []string{"ok", "action"},
+		"properties": map[string]any{
+			"ok":     map[string]any{"type": "boolean"},
+			"action": map[string]any{"type": "string", "const": action},
+		},
+	}
+}
